refactor: move all datastream sending logic into OutgoingDataStreamManager
